Refactor Instagram API integration and update analytics hooks (#464)
- Removed the deprecated  hook and replaced it with  for improved data fetching.
- Updated the  and  components to utilize the new hook, enhancing the overall structure and clarity of the code.
- Deleted the  file for the full-refresh Instagram endpoint, streamlining the API surface.
- Added error handling and improved response structures in the  endpoint to ensure better user feedback and data integrity.
- Cleaned up unused code and ensured consistent formatting across files.
