+++ convex/instagramQueries.ts
@@ -1038,3 +1038,4 @@ export const getQueueStatus = query({
     return { queueCount };
   },
 });
+
